+++ internal/dataplane/translator/translator.go
@@ -56,6 +56,7 @@ type FeatureFlags struct {
 	// CombinedServicesFromDifferentHTTPRoutes indicates whether we should combine rules from different HTTPRoutes
 	// that are sharing the same combination of backends to one Kong service.
 	CombinedServicesFromDifferentHTTPRoutes bool
+
 	// SupportRedirectPlugin indicates whether the Kong gateway supports the `redirect` plugin.
 	// This is supported starting with Kong 3.9.
 	// If `redirect` plugin is supported, we will translate the `requestRedirect` filter to `redirect` plugin
@@ -105,7 +106,17 @@ type Translator struct {
 	failuresCollector          *failures.ResourceFailuresCollector
 	translatedObjectsCollector *ObjectsCollector
 
-	clusterDomain string
+	clusterDomain      string
+	enableDrainSupport bool
+}
+
+// Config is a configuration for the Translator.
+type Config struct {
+	// EnableDrainSupport indicates whether the translator should support draining endpoints.
+	EnableDrainSupport bool
+
+	// ClusterDomain is the cluster domain used for translating Kubernetes objects.
+	ClusterDomain string
 }
 
 // NewTranslator produces a new Translator object provided a logging mechanism
@@ -116,7 +127,7 @@ func NewTranslator(
 	workspace string,
 	featureFlags FeatureFlags,
 	schemaServiceProvider SchemaServiceProvider,
-	clusterDomain string,
+	config Config,
 ) (*Translator, error) {
 	failuresCollector := failures.NewResourceFailuresCollector(logger)
 
@@ -134,7 +145,8 @@ func NewTranslator(
 		schemaServiceProvider:      schemaServiceProvider,
 		failuresCollector:          failuresCollector,
 		translatedObjectsCollector: translatedObjectsCollector,
-		clusterDomain:              clusterDomain,
+		clusterDomain:              config.ClusterDomain,
+		enableDrainSupport:         config.EnableDrainSupport,
 	}, nil
 }